refactor(settings): extract free-plan CTA into its own section
